requirements.txt
scikit-image
numpy
matplotlib
torch
torchvision
pillow
tqdm